:+1:  Show quotation list

+++ public/assets/js/quotation.js
@@ -52,23 +52,27 @@ function renderQuotationListApp() {
     if (localStorage.getItem('carrito')) {
         const carrito = JSON.parse(localStorage.getItem('carrito'));
         jQuery('#itemsList').html('');
-        jQuery.each(carrito, function(key, value){
-            
-            jQuery('#itemsList').append('<div class="row" style="padding:10px;border:1px solid #CCCCCC; border-radius:5px;">'+
-                '<div class="col-xs-12 col-md-2">'+
-                    '<img src="../'+value['img']+'" width="100%"/>'+
-                '</div>'+
-                '<div class="col-xs-12 col-md-10">'+
-                    '<b>'+value['name']+'</b>'+
-                    '<input type="hidden" name="product['+key+'][productId]" value="'+value['productId']+'"/>'+
-                    '<textarea class="form-control" name="product['+key+'][comment]" rows="2" placeholder="Agregar comentarios"></textarea>'+
-                    '<hr>'+
-                    '<button class="btn btn-danger" title="Eliminar" style="float:right" onclick=deleteItem('+key+',"APP")>'+
-                        '<i class="fa fa-trash"></i>'+
-                    '</button>'+
-                '</div>'+
-            '</div>');
-        });
+        if (carrito.length > 0){
+            jQuery.each(carrito, function(key, value){
+                
+                jQuery('#itemsList').append('<div class="row" style="padding:10px;border:1px solid #CCCCCC; border-radius:5px;">'+
+                    '<div class="col-xs-12 col-md-2">'+
+                        '<img src="../'+value['img']+'" width="100%"/>'+
+                    '</div>'+
+                    '<div class="col-xs-12 col-md-10">'+
+                        '<b>'+value['name']+'</b>'+
+                        '<input type="hidden" name="product['+key+'][productId]" value="'+value['productId']+'"/>'+
+                        '<textarea class="form-control" name="product['+key+'][comment]" rows="2" placeholder="Agregar comentarios"></textarea>'+
+                        '<hr>'+
+                        '<button class="btn btn-danger" title="Eliminar" style="float:right" onclick=deleteItem('+key+',"APP")>'+
+                            '<i class="fa fa-trash"></i>'+
+                        '</button>'+
+                    '</div>'+
+                '</div>');
+            });
+        } else {
+            jQuery('#itemsList').html('<p>No se han agregado elementos a la lista de cotización.</p>');
+        }
 
     } else {
         jQuery('#itemsList').html('<p>No se han agregado elementos a la lista de cotización.</p>');
@@ -80,4 +84,14 @@ function deleteItem(key, mode) {
     carrito.splice(key,1);
     localStorage.setItem('carrito', JSON.stringify(carrito));
     (mode === 'APP') ? renderQuotationListApp() : renderQuotationListWeb();
+}
+
+function quotationRequest() {
+    if (localStorage.getItem('carrito')) {
+        const carrito = JSON.parse(localStorage.getItem('carrito'));
+    } else {
+        jQuery('#alerta').html('<div class="alert alert-warning">'+
+            '<p>No hay elementos en la lista de cotización.</p>'+
+        '</div>')
+    }
 }
